# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Contacts API",
    "description": "RESTful API for managing contacts in MongoDB",
    "version": "1.0.0"
  },
  "host": "web-services-app.onrender.com",
  "basePath": "/",
  "schemes": ["https"],
  "paths": {
    "/": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/contacts/": {
      "get": {
        "tags": ["Contacts"],
        "summary": "Get all contacts",
        "description": "Returns an array of every contact in the database.",
        "responses": {
          "200": {
            "description": "Array of contact objects",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Contact"
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Contacts"],
        "summary": "Create a new contact",
        "description": "Adds a new contact document to the database.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Contact data",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ContactInput"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Contact created — returns the new document id"
          },
          "400": {
            "description": "Missing required fields"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/contacts/single": {
      "get": {
        "tags": ["Contacts"],
        "summary": "Get a contact by ID",
        "description": "Returns a single contact matching the provided MongoDB _id query parameter.",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "MongoDB ObjectId of the contact",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "A single contact object",
            "schema": {
              "$ref": "#/definitions/Contact"
            }
          },
          "404": {
            "description": "Contact not found"
          }
        }
      }
    },
    "/contacts/{id}": {
      "put": {
        "tags": ["Contacts"],
        "summary": "Update a contact by ID",
        "description": "Replaces the fields of an existing contact identified by its MongoDB _id.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "MongoDB ObjectId of the contact to update"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Updated contact data",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ContactInput"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Contact updated successfully"
          },
          "400": {
            "description": "Invalid id or missing required fields"
          },
          "404": {
            "description": "Contact not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "tags": ["Contacts"],
        "summary": "Delete a contact by ID",
        "description": "Permanently removes the contact with the given MongoDB _id from the database.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "MongoDB ObjectId of the contact to delete"
          }
        ],
        "responses": {
          "200": {
            "description": "Contact deleted successfully"
          },
          "400": {
            "description": "Invalid id format"
          },
          "404": {
            "description": "Contact not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  },
  "definitions": {
    "Contact": {
      "type": "object",
      "properties": {
        "_id": { "type": "string", "example": "64b1f2c3e4b0a1b2c3d4e5f6" },
        "firstName": { "type": "string", "example": "Jane" },
        "lastName": { "type": "string", "example": "Doe" },
        "email": { "type": "string", "example": "jane@example.com" },
        "favoriteColor": { "type": "string", "example": "blue" },
        "birthday": { "type": "string", "example": "1990-01-01" }
      }
    },
    "ContactInput": {
      "type": "object",
      "required": [
        "firstName",
        "lastName",
        "email",
        "favoriteColor",
        "birthday"
      ],
      "properties": {
        "firstName": { "type": "string", "example": "Jane" },
        "lastName": { "type": "string", "example": "Doe" },
        "email": { "type": "string", "example": "jane@example.com" },
        "favoriteColor": { "type": "string", "example": "blue" },
        "birthday": { "type": "string", "example": "1990-01-01" }
      }
    }
  }
}